pyproject.toml
[project]
name = "pdfding"
# do not change the version line position as it will break the automated update after a release
version = "1.4.1"
description = "Selfhosted PDF manager, viewer and editor offering a seamless user experience on multiple devices."
authors = [
  { name = "mrmn2" },
]
readme = "README.md"
license = "AGPL-3.0-only"
license-files = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.11 <4.0"
django = "==5.2.8"
django-allauth = { extras = ["socialaccount"], version = "==65.13.0" }
django-cleanup = "==9.0.0"
django-htmx = "==1.26.0"
django-rest-knox = "==5.0.2"
djangorestframework = "==3.16.1"
gunicorn = "==23.0.0"
huey = "==2.5.4"
Markdown = "==3.10"
minio = "==7.2.18"
nh3 = "==0.3.2"
psycopg2-binary = "==2.9.11"
pypdfium2 = "==5.0.0"
pypdf = "==6.1.3"
python-magic = "==0.4.27"
qrcode = { extras = ["pil"], version = "==8.2" }
RapidFuzz = "==3.14.3"
ruamel-yaml = "==0.18.16"
supervisor = "==4.3.0"
whitenoise = { extras = ["brotli"], version = "==6.11.0" }

[tool.poetry.group.dev.dependencies]
bandit = "==1.9.2"
black = "==25.12.0"
flake8 = "==7.3.0"
pytest = "==9.0.2"
pytest-cov = "==7.0.0"
pytest-django = "==4.11.1"

[tool.poetry.group.e2e.dependencies]
pytest-playwright = "==0.7.2"

[tool.black]
line-length = 120
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "pdfding.core.settings"

[tool.bandit]
exclude_dirs = ["tests", "path/to/file"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
